Breaking Mongo config out into config.json
This makes it easier to develop the app with a local Mongo server that isn't
hosted on localhost.  Otherwise, I have to change "localhost" in the .js file
and then be very careful not to commit that particular line change.  This way, I
can simply edit config.json and never worry about committing that file, rather
than a single line in the source.